+++ DRN.Framework.Testing/Contexts/MethodContext.cs
@@ -1,3 +1,4 @@
+using DRN.Framework.SharedKernel;
 using DRN.Framework.Testing.Extensions;
 using DRN.Framework.Utils.DependencyInjection;
 using DRN.Framework.Utils.Extensions;
@@ -7,6 +8,9 @@ namespace DRN.Framework.Testing.Contexts;
 
 public class MethodContext(MethodInfo testMethod)
 {
+    private readonly object _tempPathLock = new();
+    private string? _tempPath;
+
     public MethodInfo TestMethod { get; } = testMethod;
     public IReadOnlyList<object> Data { get; private set; } = [];
     public IReadOnlyList<SubstitutePair> SubstitutePairs { get; private set; } = [];
@@ -19,6 +23,41 @@ public class MethodContext(MethodInfo testMethod)
         var testFolder = Path.Combine(Path.GetDirectoryName(testClass.Assembly.Location) ?? "", relativePathToTest);
 
         return testFolder;
+    }
+
+    public string GetTempPath()
+    {
+        if (_tempPath != null)
+            return _tempPath;
+
+        lock (_tempPathLock)
+            return _tempPath ??= CreateTempPath();
+    }
+
+    internal void DeleteTempPath()
+    {
+        lock (_tempPathLock)
+        {
+            if (string.IsNullOrWhiteSpace(_tempPath))
+                return;
+
+            if (Directory.Exists(_tempPath))
+                Directory.Delete(_tempPath, true);
+
+            _tempPath = null;
+        }
+    }
+
+    private string CreateTempPath()
+    {
+        var testClass = TestMethod.ReflectedType;
+        var testClassPath = (testClass?.FullName ?? TestMethod.Name).Replace('.', Path.DirectorySeparatorChar);
+        var tempPath = Path.Combine(AppConstants.TempPath, testClassPath, TestMethod.Name, Guid.NewGuid().ToString("N"));
+
+        if (Directory.Exists(tempPath)) Directory.Delete(tempPath, true);
+        Directory.CreateDirectory(tempPath);
+
+        return tempPath;
     }
 
     internal void SetTestData(object[] dataRow)